requirements.txt
numpy==1.24.3
Pillow==9.4.0
Pillow==10.1.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
wandb==0.15.11
